feat: Implement useTextNode composable for improved state management and reactivity in TextNode component
