+++ frontend/src/composables/useTextNode.js
@@ -1,0 +1,85 @@
+import { ref, computed, onMounted } from 'vue'
+import { useVueFlow } from '@vue-flow/core'
+
+export default function useTextNode(props, emit) {
+  const { getEdges, findNode } = useVueFlow()
+  
+  // Custom style for handling resizes
+  const customStyle = ref({})
+  
+  // Track whether the node is hovered (to show/hide the resize handles)
+  const isHovered = ref(false)
+  
+  const resizeHandleStyle = computed(() => ({
+    visibility: isHovered.value ? 'visible' : 'hidden',
+    width: '12px',
+    height: '12px',
+  }))
+  
+  // Computed property for two-way binding of the text input
+  const text = computed({
+    get: () => props.data.inputs.text,
+    set: (value) => {
+      props.data.inputs.text = value
+      updateNodeData()
+    }
+  })
+  
+  onMounted(() => {
+    if (!props.data.run) {
+      props.data.run = run
+    }
+  })
+  
+  // Execute the node's logic
+  async function run() {
+    const connectedSources = getEdges.value
+      .filter((edge) => edge.target === props.id)
+      .map((edge) => edge.source)
+  
+    if (connectedSources.length > 0) {
+      const sourceNode = findNode(connectedSources[0])
+      if (sourceNode && sourceNode.data.outputs.result) {
+        props.data.inputs.text = sourceNode.data.outputs.result.output
+      }
+    }
+  
+    // Set the output equal to the current text input
+    props.data.outputs = {
+      result: {
+        output: text.value
+      }
+    }
+  }
+  
+  // Emit updated node data back to VueFlow
+  function updateNodeData() {
+    const updatedData = {
+      ...props.data,
+      inputs: { text: text.value },
+      outputs: props.data.outputs
+    }
+    emit('update:data', { id: props.id, data: updatedData })
+  }
+  
+  // Handle the resize event to update the node dimensions
+  const onResize = (event) => {
+    customStyle.value.width = `${event.width}px`
+    customStyle.value.height = `${event.height}px`
+    // Also update the node's style data so it persists
+    props.data.style.width = `${event.width}px`
+    props.data.style.height = `${event.height}px`
+    updateNodeData()
+    emit('resize', { id: props.id, width: event.width, height: event.height })
+  }
+  
+  return {
+    text,
+    customStyle,
+    isHovered,
+    resizeHandleStyle,
+    updateNodeData,
+    onResize,
+    run
+  }
+}
